+++ doc/conf.py
@@ -247,7 +247,8 @@ intersphinx_mapping = {
     'data-quality-report': (
         'https://docs.ligo.org/detchar/data-quality-report/', None),
     'gwpy': ('https://gwpy.github.io/docs/stable/', None),
-    'ligo.skymap': ('https://leo-singer.docs.ligo.org/ligo.skymap/', None)
+    'ligo.skymap': ('https://leo-singer.docs.ligo.org/ligo.skymap/', None),
+    'userguide': ('https://emfollow.docs.ligo.org/userguide/', None)
 }